backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.5.2
requests==2.31.0
python-multipart>=0.0.6
astroid>=3.0.1
python-dotenv==1.0.1
openai>=1.3.0
pydantic-settings>=2.9.1
pydantic-extra-types>=2.10.4
transformers>=4.36.0
torch>=2.1.0 